requirements.txt
asgiref==3.5.2
attrs==21.4.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.5
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
isort==5.10.1
packaging==21.3
pluggy==1.0.0
psycopg2==2.8.6
py==1.11.0
pycodestyle==2.8.0
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.2
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
urllib3==1.26.9
